backend/requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
Django==4.1
djangorestframework==3.13.1
frozenlist==1.3.1
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
multidict==6.0.2
pytz==2022.1
requests==2.28.1
rfc3986==1.5.0
sqlparse==0.4.2
urllib3==1.26.11
yarl==1.8.1
